requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
python-dotenv
pymupdf
pillow
